+++ Tests/CubiomesKitTests/MapKit Interop/MinecraftMapViewContentTests.swift
@@ -20,7 +20,7 @@ struct MinecraftMapViewContentTests {
         #expect(view.annotations.isEmpty)
 
         let contents: [AnyMinecraftMapContent] = buildMinecraftMapContent {
-            Marker(location: .zero, title: "Origin")
+            Marker(location: .zero, title: "Origin", id: "sample")
         }
         view.resyncMapContentIfNeeded(contents)
         #expect(view.annotations.count == 1)
@@ -88,13 +88,42 @@ struct MinecraftMapViewContentTests {
         }
 
         #expect(
-            playerMarker.isEqual(
-                to: MinecraftMapPlayerMarkerAnnotation(
-                    name: "lweiss",
-                    playerUUID: id,
-                    location: CGPoint(x: 1847, y: 1963)
-                )
-            )
+            playerMarker.model == PlayerMarker(location: CGPoint(x: 1847, y: 1963), name: "lweiss", playerUUID: id)
+        )
+    }
+
+    @Test(.tags(.mapkit))
+    func mapViewUpdatesMarkerContentInPlace() throws {
+        let view = MinecraftMapView(world: world, frame: .zero)
+        #expect(view.annotations.isEmpty)
+
+        let id = UUID()
+        let contents: [AnyMinecraftMapContent] = buildMinecraftMapContent {
+            Marker(location: .zero, title: "Spawn", id: id)
+        }
+        view.resyncMapContentIfNeeded(contents)
+
+        guard let marker = view.annotations.first as? MinecraftMapMarkerAnnotation else {
+            Issue.record("The annotation isn't a marker, or no markers exist.")
+            return
+        }
+
+        #expect(
+            marker.model == Marker(location: .zero, title: "Spawn", id: id)
+        )
+
+        let newContents: [AnyMinecraftMapContent] = buildMinecraftMapContent {
+            Marker(location: CGPoint(x: 1847, y: 1847), title: "Spawn", id: id)
+        }
+        view.resyncMapContentIfNeeded(newContents)
+
+        guard let playerMarker = view.annotations.first as? MinecraftMapMarkerAnnotation else {
+            Issue.record("The annotation isn't a marker, or no markers exist.")
+            return
+        }
+
+        #expect(
+            playerMarker.model == Marker(location: CGPoint(x: 1847, y: 1847), title: "Spawn", id: id)
         )
     }
 }